pyproject.toml
[tool.poetry]
name = "web-scraper"
version = "0.1.0"
description = ""
authors = ["Marius Bause <mariusbause@gmx.de>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.13"
streamlit = "^1.41.1"
beautifulsoup4 = "^4.12.3"
scrape = "^0.11.3"
scraping = "^0.0.3"
langchain-ollama = "^0.2.2"
scrapy = "^2.12.0"
selenium = "^4.27.1"
python-dotenv = "^1.0.1"
parse = "^1.20.2"
llama-cpp-python = "^0.3.5"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
